fifth commit

+++ course app/Models/student_main_page.cs
@@ -11,6 +11,6 @@ namespace course_app.Models
         public string section_date { get; set; }
         public string instructor_name { get; set; }
         public int section_id { get; set; }
-        public int grade { get; set; }
+        public double grade { get; set; }
     }
 }
